% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_slice.R
\name{layer_slice}
\alias{layer_slice}
\alias{layer_slice.multimixMN_fit}
\title{Extract a single layer from a multilayer MixMashNet object}
\usage{
layer_slice(object, ...)

\method{layer_slice}{multimixMN_fit}(object, layer, ...)
}
\arguments{
\item{object}{An object of class \code{"multimixMN_fit"} returned by
\code{multimixMN()}.}

\item{...}{Further arguments passed to methods.}

\item{layer}{Character string giving the layer to extract.}
}
\value{
An object of class \code{"mixMN_fit"} corresponding to the
selected layer.
}
\description{
Extracts one layer from a fitted multilayer \code{multimixMN_fit} object
returned by \code{multimixMN()}.

The selected layer is returned as the corresponding single layer
\code{mixMN_fit} object stored in \code{layer_fits}.
}
