% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes.R
\docType{data}
\name{nhanes}
\alias{nhanes}
\title{NHANES dataset}
\format{
A data frame with 2759 rows and 29 variables:
\describe{
  \item{TotChol}{Total Cholesterol (numeric).}
  \item{HDL}{High-density lipoprotein cholesterol (numeric).}
  \item{Creatinine}{Creatinine (numeric).}
  \item{UricAcid}{Uric acid (numeric).}
  \item{ALT}{Alanine aminotransferase (numeric).}
  \item{AST}{Aspartate aminotransferase (numeric).}
  \item{GGT}{Gamma-glutamyl transferase (numeric).}
  \item{Bilirubin}{Bilirubin (numeric).}
  \item{Albumin}{Albumin (numeric).}
  \item{TotProtein}{Total protein (numeric).}
  \item{HbA1c}{Glycated hemoglobin (numeric).}
  \item{hsCRP}{High-sensitivity C-reactive protein (numeric).}
  \item{BMI}{Body mass index (numeric).}
  \item{Waist}{waist circumference (numeric).}
  \item{Height}{Height (numeric).}
  \item{ArmCirc}{Arm circumference (numeric).}
  \item{HipCirc}{Hip circumference (numeric).}
  \item{LegLength}{Leg length (numeric).}
  \item{ArmLength}{Arm Length (numeric).}
  \item{TroubleSleep}{Trouble Sleep with 0=no and 1=yes.}
  \item{PhysicalActivity}{Physical Activity with 0=no and 1=yes.}
  \item{Smoke}{Smoking with 0=no and 1=yes.}
  \item{Drug}{Drug use with 0=no and 1=yes.}
  \item{Diet}{Dietary quality with 1=Poor, 2=Fair, 3=Good, 4=Very good, 5=Excellent.}
  \item{Alcohol}{Alcohol consumption with 0=no and 1=yes.}
  \item{Work}{Employment status (1–4: working; employed but absent; seeking work; not working).}
  \item{MonInc}{Monthly income category (1–12), where higher values indicate higher income.}
  \item{Gender}{Gender with 0=male and 1=female.}
  \item{Age}{Age (numeric).}
}
}
\usage{
data(nhanes)
}
\description{
Example dataset used in MixMashNet examples. This dataset contains 29 variables
derived from the National Health and Nutrition Examination Survey (NHANES)
}
\references{
Centers for Disease Control and Prevention (CDC) & National Center for Health
Statistics (NCHS). (2020).
National Health and Nutrition Examination Survey data.
\url{https://www.cdc.gov/nchs/nhanes/}
}
