% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bridge_communities.R
\name{print.bridge_profiles}
\alias{print.bridge_profiles}
\title{Print method for objects of class \code{"bridge_profiles"}}
\usage{
\method{print}{bridge_profiles}(
  x,
  statistic = c("bridge_strength", "bridge_ei1", "bridge_ei2", "bridge_closeness",
    "bridge_betweenness"),
  digits = 3,
  ...
)
}
\arguments{
\item{x}{An object of class \code{"bridge_profiles"}.}

\item{statistic}{Character string indicating which bridge profile to print.
If missing, all available profiles are printed.}

\item{digits}{Number of decimal digits used when printing numeric results.}

\item{...}{Further arguments passed to or from other methods (currently unused).}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print method for objects of class \code{"bridge_profiles"}
}
