% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mixMN_fit}
\alias{print.mixMN_fit}
\title{Print method for single layer MixMashNet objects}
\usage{
\method{print}{mixMN_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"mixMN_fit"}.}

\item{...}{Additional arguments.}
}
\value{
The input object \code{x}, returned invisibly.
}
\description{
Compact textual summary for objects returned by \code{mixMN()}.
}
\seealso{
\code{\link{mixMN}}
}
