% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.multimixMN_fit}
\alias{print.multimixMN_fit}
\title{Print method for multilayer MixMashNet objects}
\usage{
\method{print}{multimixMN_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"multimixMN_fit"}.}

\item{...}{Additional arguments.}
}
\value{
The input object \code{x}, returned invisibly.
}
\description{
Compact textual summary for objects returned by \code{multimixMN()}.
}
\seealso{
\code{\link{multimixMN}}
}
