% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmashnet-summary.R
\name{print.summary.mixmashnet}
\alias{print.summary.mixmashnet}
\title{Print method for \code{"summary.mixmashnet"} objects}
\usage{
\method{print}{summary.mixmashnet}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{"summary.mixmashnet"} object as returned by
\code{summary.mixmashnet()}.}

\item{digits}{Number of digits to print.}

\item{...}{Further arguments passed to or from other methods (currently unused).}
}
\description{
Print method for \code{"summary.mixmashnet"} objects
}
